fix(config): share community base resolver
